Test working and add logging
Adding some logging to show which devices are being found every period
of time.

I've tested the working-ness of it by creating a super simple C program
which immitates hcitool and put it in my path. Here it is:

```c
\#include <stdio.h>
int main(void) {
    printf("Searching...\\n\\tA   B\\n\\tC   D");
    return 0;
}
```

With this I was able to test whether the genserver was running every ten
seconds and making the call to hcitool, and parsing it correctly.