Give a blank map as initial input to GenServer
This is the proper arg for the job as it allows the genserver to start
with an empty map, which was my original intention. This solves an issue
where, after the server died (from timing out, or initial start), the
rendered page would look odd as the arg passed to it as `@devices` was
actually the empty map. Now it should render correctly and if there are
no devices, there will be no output